Add unittest which passes an inline template into an included template.
The TemplateBase.WriteTo used a different writer for IEncodedString than
for other types. In most situations the writer argument contains the same
writer as the _context.CurrentWriter, but not when passing an inline
template into an included template. As a result the
output was in the wrong order. This bug was fixed in
67e2f20a036fa52d04e9fd8228cbbf2fd94e8a8b

Conflicts:
	src/source/RazorEngine.Core/Templating/TemplateBase.cs
	src/test/Test.RazorEngine.Core/TemplateBaseTestFixture.cs
	src/test/Test.RazorEngine.Core/Test.RazorEngine.Core.csproj