Bugfix. Passing inline templates into included templates leads to wrong order in output.
The TemplateBase.WriteTo used a different writer for IEncodedString than
for other types. In most situations the writer argument contains the same
writer as the _context.CurrentWriter, but not when passing an inline
template in a parent template into a child template. As a result the
output was in the wrong order.