Added GetAllTemplatKeys()
Note: Please don't merge this commit as-is, because it's meant just to give you a hint about the possible addition.
This method helps to enumerate all the embedded resources in a given assembly, identified by RootType. A common scenario for EmbeddedResourceTemplateManager usage would be to enumerate all the embedded resources and Resolve/Cache them with Razor Engine, for future use. Not sure if the method needs to be moved to the interface as well. Without this method, it's a bit difficult to use EmbeddedResourceTemplateManager, since users will have to write this method themselves and will have to pay attention to exact key naming (without prefix and suffix).